Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,7 +3,7 @@
 int main(){
     float radius;
     scanf("%f", &radius);
-    int pi= 3.14;
+    float pi= 3.14;
 
     print("Area: %.2f", pi*radius*radius;)
 }
